Make a base class for row and move all animations there
