+++ lib/features/drug_list/bloc/drug_list_bloc.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
+import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:my_drugs/data_access/data_access.dart';
 import 'package:my_drugs/features/drug_list/drug_list_item.dart';
@@ -14,8 +15,11 @@ enum ScreenMode { normal, edit }
 
 class DrugListBloc extends Bloc<DrugListEvent, DrugListState> {
   final AbstractDrugRepository _repository;
-  List<Drug> _drugs;
+  final List<Drug> _drugs;
   List<String> _selectedDrugsIds = [];
+
+  /// Items that are currently displaying on DrugListScreen.
+  List<DrugListItem> _listItems;
 
   final DateFormat _dateFormat = DateFormat('MMM yyyy');
   ScreenMode _screenMode = ScreenMode.normal;
@@ -37,23 +41,30 @@ class DrugListBloc extends Bloc<DrugListEvent, DrugListState> {
   @override
   DrugListState get initialState => _buildState();
 
-  DrugListState _buildState() => _drugs.isEmpty
-      ? DrugListEmpty()
-      : DrugListLoaded(
-          _screenMode,
-          _buildListItems(),
-          '${_drugs.length} items',
-          '${_selectedDrugsIds.length} selected',
-          _selectedDrugsIds.isNotEmpty,
-        );
+  DrugListState _buildState() {
+    if (_drugs.isEmpty) {
+      return DrugListEmpty();
+    } else {
+      if (_listItems == null) {
+        _listItems = _buildListItems(_drugs);
+      }
+      return DrugListLoaded(
+        _screenMode,
+        _listItems,
+        '${_drugs.length} items',
+        '${_selectedDrugsIds.length} selected',
+        _selectedDrugsIds.isNotEmpty,
+      );
+    }
+  }
 
-  List<DrugListItem> _buildListItems() {
+  List<DrugListItem> _buildListItems(List<Drug> drugs) {
     List<DrugListItem> result = [];
     List<Drug> expired = [];
     List<Drug> notExpired = [];
     final now = DateTime.now();
     final firstDayOfCurrentMonth = DateTime(now.year, now.month, 1);
-    _drugs.forEach((e) {
+    drugs.forEach((e) {
       if (e.expiresOn.compareTo(firstDayOfCurrentMonth) > 0) {
         notExpired.add(e);
       } else {
@@ -62,11 +73,15 @@ class DrugListBloc extends Bloc<DrugListEvent, DrugListState> {
     });
     if (expired.isNotEmpty) {
       result.add(
-        DrugHeadingItem('EXPIRED'),
+        DrugHeadingItem(
+          GlobalKey(),
+          'EXPIRED',
+        ),
       );
       result.addAll(
         expired.map(
           (e) => DrugItem(
+            GlobalKey(),
             e.id,
             e.name,
             _dateFormat.format(e.expiresOn),
@@ -76,11 +91,15 @@ class DrugListBloc extends Bloc<DrugListEvent, DrugListState> {
     }
     if (notExpired.isNotEmpty) {
       result.add(
-        DrugHeadingItem('NOT EXPIRED'),
+        DrugHeadingItem(
+          GlobalKey(),
+          'NOT EXPIRED',
+        ),
       );
       result.addAll(
         notExpired.map(
           (e) => DrugItem(
+            GlobalKey(),
             e.id,
             e.name,
             _dateFormat.format(e.expiresOn),
